Update jsRecordCollection.js
Their labeling of the parameters gave me some grief, but after I understood the problem it was easier. I learned about adding an empty array properly as well as adding a parameter to that array. Using hasOwnProperty was helpful, i'll have a better chance of adding it to my coding dictionary now that i've worked with it. Also, punctuation. 'tracks'. very key.