Create jsIterateForLoop.js
Prior, while loops work as long as the condition is true. With for loops, they work as "for this many times". For loops are 3 part, seperated by a semi-colon . Initialization of the variable; Condition; and Final-Expression (meaning something that happens after the loop has finished, usually an iteration factor to move along the conditional statement/your count. Interesting stuff, looking forward to more practice to add this firmly* to my coding lexicon.