Create jsReplaceLoopsRecursion.js
Invested a lot into this exercise to get more insight on recursions. What seems to be the case is that recursions decrement down to iterate through some list/array. In the original function, we had a for loop that read:

for (var i = 0; i < n; i++)

which is incrementing.  The equivalent of this functions 'base case' is the conditional statement (i<n).

Looking at (i<n), I remembered that we use the variable i we initialized to use as a counter. This counter uses bracket notation which is zero indexed, meaning that if there is some array with length n, it will never reach 'n'. So, when working with zero indexing, we have to write it out like (n-1) if we want the function to run/get all the values.

The example problem also gave me some grief because they made the statement:

"However, notice that multiply(arr, n) == multiply(arr, n - 1) * arr[n - 1]. "

I took pen to paper to work through how this worked, then figured the utility of it out, all step by step. I saw n-1 and though- "oh, it's the issue of indexing I was working with earlier. How did that go?" Watched a few videos, googled a bit, and then had a better grasp of what I was seeing.  arr[n-1] pulls the last value off of the array, then multiply adds to the stack by calling itself with a decremented n value, which will use arr[n-1] again. When it all resolves, everything multiploes together.

I then asked myself, well, how will I know I can do this? Obvious answer is when I see a for loop. But the whole "however, notice that..." did not come very naturally, so I looked a bit more at the issue.

Anything being incremented/decrimented, from what it looks like, can be rewritten using recursion. All that's needed is the ability to write off to 'the side' one of the 'n' values from the array, then be sure to call the function again, this time, decrimenting it.

Also, don't forget the base case.