Create jsBasicJavaScriptProfileLookup.js
This one stumped me for a bit. I don't think freeCodeCamp allows you to deviate much. I'll spend more time tinkering with this one.  I was surprised that:

if (contacts[a].hasOwnProperty(name) && contacts[a].hasOwnProperty(prop)) {return contacts[a][prop];

did not pass any tests. Ill post on a forum or slack to get feedback.