Create jsUseRecursionCreateCountdown.js
base case
recursion line
the push of this instances parameter/number

initialize recursion line and equal it to a variable to return---
---if it goes after the recursion line, whatever is at  the end will be---
appended first

You want to add and push the thing that's repeating.
You really do need to have the const equal the recursion, you need to push the recursion
after it resolves AND return the const
Acting on the const twice here, pushing onto the const & returning it (to the new box/funct)
Pushing the countArrays value of a resolved recursion is so neat and tidy, then returning
it is just the same, neat and tidy. When we're done and the base line condition is reached,
we give it an empty set to return as the recursions value. Then, we just push onto it,
then return the new pushed product. That's pretty freakin' cool.

The question to have asked myself is: "How do I reuse that recursion line?"
By attaching it to a variable(const), which we also conveniently need.
We work with the variable using methods and by returning it (to the new recursion box)
The recursion structure itself is a neat mechanis of it's own.

Returning the empty slate at the end of the countdown is pretty smart, too.

Need a var/const? Yes
Use it to build something? Yes
Is that something dependent on iterative input? Yes
Consider declaring it's value as the recursion line. We're afforded an opportunity to build
an (array) using methods and return statements to the next box(recursion function).