Create jsRecursionCreateRangeNumbers.js
Same build as prior use of recursion to build an array. When a condition is met, return an empty set.

Initialize a variable, and set it equal to the recursion line because we are going to be working with the recursions output. Note the parameters used, there are two here. With there being two, we need to have two parameters when we call the function we are in. We want to decrimate endNum, so we subtract in the parameter so the argument passed will be one less. What we want to repeat is the adding of the number to the list. This will not happen until the condition is met, we're given an open set by return, and at this point we'll be at the lowest possible value of endNum.

The process will run, check the if statement, then run else and start stacking rangeOfNumbers until we've decrimented one less than the start num, with a stack that's already dipped/went lower, meaning it will resolve the lower numbers first.  The process will go:

[] (push endNum(lowest)) return rangeArray

(push endNum(lower)) return
rangeArray

until we work ourway up the function stack!