Create jsArrowFunctionsConciseAnonFunctions.js
Exercise problem introduced a concise way to write anonymous functions. We sometimes dont have to name functions because we don't use them anywhere else, especially in situations where we pass them as arguments to other functions.

Here's an example of such a situation:

const funkyFunk = function() {
const varusVar= "value";
return varusVar;
}

Here's the syntatic sugar:

const funkyFunk = () => {
const varusVar = "value";
return varusVar;
}

Essentially it saves us from having to write out the name of the single serving function, and instead drop a () in it's stead along with an arrow =>.

The other utility of it is when there's only a return value within the function being assigned to a variable, we get to drop the keyword 'return' and the brackets surrounding the code, making it a one line statement:

const func = () => "value";