Create jsArrowFuncParam.js
A lot of helpful info from this example problem. With arrow function notation we are just as able to pass arguments/parameters into the functions.

If there's just one parameter, we don't even have to use parenthesis.

The exercise would not allow for a pass as long as I did not change the var keyword to const. I realized this was the case because for one time use functions meant to pass information likely to other functions, we did not want to alter the information stored with in as a global var scope could do. We would ideally like this information to be read-only, too, so the const key word is ideal.