Create jsES6DestructuringAssignmentRestParameterReassignArrayElements.js
Very straight forward, used the deconstruction assignment to pull the first two values from the array and assign them to a & b, just like in the example. Then, I used the rest parameter to gather the left over array.

The point was made that we cannot use the rest parameter anywhere else but as the last element in the parameter. We can only catch what's left of the array (the rest of it).