Create jsES6DestructuringAssignemtPassObjectAsParameter.js
We are learning here how to pass an object as a functions parameter.

We are also seeing how to destructure an object in a function argument.

const profileUpdate = (profileData) => { name, age, nationality, location } = profileData;
//use variables
}

This destructures the object sent into the function so we can use the property keys as variables.

We can also do this in-place:

const profileUpdate = ({ name,age,nationality,location}) => {
}
Meaning that when the argument, the object, is passed through, we can destruct in by virtue of what we have written in the parameter