Update jsES6CreateStringsTemplateLiterals.js
return const failureItems with all values from result objects property failure

spread?
or
destructure?

structure of : '<li class="text-waring">no-var</li>'
can i access using zero indexing each value I need? yes!
can I return those values to the failureItems? Yes! push?
I'd need to push each of them as a newline* onto the const

Two things I could work on:
1) realizing that I don't need to make a variables as much as I do. In this problem, I could have just inserted
the string with the `${}` template literal substitution in it
2) I read the prompt initially and didn't ascertain that it's a function that takes an array name, not a specific set of
keys as I thought (because of how the problem was worded). Make a point to read every line of code thoughtfully,
asking myself why i am doing this or that.

extra) The syntax for the desired output was done in my original function, but for whatever reason
it's preferable to the check that I don't have '', anywhere in it. I thought I was slick making a template literals
functionality to make a newline.