Create jsES6ClassSyntaxDefineConstructor.js
I've never seen a constructor function before, so I looked it up on w3. I took some time to familiarize myself with it so I wouldn't have to be checking how to write out its structure while solving the problem.

The one hiccup was I took the liberty of naming the key whatever I wanted (it was 'veggie'). The test didn't pass, so I read through each line of my code, checked it against the example, then read passed where we're told to edit. As it turns out, the key name used in the test was ____.name !