Create jsES6GettersSettersControlAccessToObject.js
Challenging to remember the syntax after taking a day off for construction functions & the ES6 shortcut.

The big take away here is that getters and setters act as internal functions of a class/constructor function to access and change (get/set) values. Javascript gets close to using private  variables by prepending an '_' to a variable, but it's just in name.